build: patterns to bypass ghostscript version detection.
Do not try to follow up on the installed ghostscript (otherwise we'll
always have to fix as MSYS2 packages gets updated). Instead, just
install whatever is in the /share/ghostscript/ path (assuming there is
data for a single version, which should be the case, especially on a
resetted CI build) and excluding the /doc/ folder.